Rename JBrowse.js to JBrowseView.js
