+++ src/components/JBrowse/JBrowseView.js
